# src/main/resources/docs/api.yaml
openapi: 3.1.0
info:
  title: HFF Data Studio API
  description: HFF Data Intelligence Platform - CATalyst API
  version: 1.0.0
  contact:
    name: Data Analyst Service - HFF
    email: data-analyst@hff.re
  license:
    name: HFF Internal Use - Restricted Distribution
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://catalyst.hff.re/api/v1
    description: Production server
...
paths:
...
  /admin/users:
    get:
      tags: [Users]
      summary: Get all users
      operationId: getAllUsers
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated list of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageUserResponse"
    post:
      tags: [Users]
      summary: Create a new user
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreationRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
  /admin/users/active:
    get:
      tags: [Users]
      summary: Get active users
      operationId: getActiveUsers
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated list of active users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageUserResponse"
  /admin/users/{userId}:
    get:
      tags: [Users]
      summary: Get user by ID
      operationId: getUserById
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
    put:
      tags: [Users]
      summary: Update user
      operationId: updateUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
    delete:
      tags: [Users]
      summary: Deactivate user
      operationId: deactivateUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: User deactivated
  /admin/users/{userId}/activate:
    patch:
      tags: [Users]
      summary: Activate user
      operationId: activateUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: User activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
  /admin/users/{userId}/password:
    patch:
      tags: [Users]
      summary: Change user password
      operationId: changePassword
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "204":
          description: Password changed
  /admin/users/email/{email}:
    get:
      tags: [Users]
      summary: Get user by email
      operationId: getUserByEmail
      parameters:
        - name: email
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
  /admin/users/category/{categoryId}:
    get:
      tags: [Users]
      summary: Get users by category
      operationId: getUsersByCategory
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated list of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageUserResponse"
  /admin/users/search:
    get:
      tags: [Users]
      summary: Search users
      operationId: searchUsers
      parameters:
        - name: searchTerm
          in: query
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageUserResponse"
...
  /admin/users/statistics/monthly:
    get:
      tags: [Users]
      summary: Get users grouped by month
      operationId: getUsersByMonth
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Monthly user statistics
          content:
            application/json:
              schema:
                type: array
                items:
                  type: array

  # ==================== CATEGORIES ====================
...
  /projects/user/{userId}:
    get:
      tags: [Projects]
      summary: Get projects by user
      operationId: getProjectsByUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated list of projects
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageProjectResponse"
...
  /admin/audit/user/{userId}:
    get:
      tags: [Audit]
      summary: Get audit logs by user
      operationId: getAuditLogsByUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated audit logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageAuditLogResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        lastName:
          type: string
        firstName:
          type: string
        email:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        lastLogin:
          type: string
          format: date-time
        category:
          $ref: "#/components/schemas/UserCategoryResponse"
    UserCreationRequest:
      type: object
      properties:
        lastName:
          type: string
        firstName:
          type: string
        email:
          type: string
        password:
          type: string
        categoryId:
          type: string
          format: uuid
      required: [lastName, firstName, email, password, categoryId]
    UserUpdateRequest:
      type: object
      properties:
        lastName:
          type: string
        firstName:
          type: string
        email:
          type: string
        isActive:
          type: boolean
        categoryId:
          type: string
          format: uuid
        newPassword:
          type: string
    ChangePasswordRequest:
      type: object
      properties:
        newPassword:
          type: string
      required: [newPassword]
    UserCategoryResponse:
      type: object
      properties:
        categoryId:
          type: string
          format: uuid
        label:
          type: string
        description:
          type: string
        accessLevel:
          type: integer
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/PermissionResponse"
    PermissionResponse:
      type: object
      properties:
        permissionId:
          type: string
          format: uuid
        code:
          type: string
        label:
          type: string
        module:
          type: string
...
    ProjectStatus:
      type: string
      enum: [IN_PROGRESS, COMPLETED, ARCHIVED]
    ProjectResponse:
      type: object
      properties:
        projectId:
          type: string
          format: uuid
        projectName:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        status:
          $ref: "#/components/schemas/ProjectStatus"
        creatorUserId:
          type: string
          format: uuid
        creatorFullName:
          type: string
        fileCount:
          type: integer
          format: int64
        datasetCount:
          type: integer
          format: int64
...
    AuditLogResponse:
      type: object
      properties:
        logId:
          type: string
          format: uuid
        action:
          type: string
        concernedEntity:
          type: string
        entityId:
          type: string
          format: uuid
        actionDate:
          type: string
          format: date-time
        ipAddress:
          type: string
        userId:
          type: string
          format: uuid
        userFullName:
          type: string
        userEmail:
          type: string

    # ===== PAGINATION =====
    PageResponse:
      type: object
      properties:
        content:
          type: array
        pageNumber:
          type: integer
        pageSize:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        isFirst:
          type: boolean
        isLast:
          type: boolean
        hasNext:
          type: boolean
        hasPrevious:
          type: boolean
    PageUserResponse:
      allOf:
        - $ref: "#/components/schemas/PageResponse"
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: "#/components/schemas/UserResponse"
    PageProjectResponse:
      allOf:
        - $ref: "#/components/schemas/PageResponse"
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: "#/components/schemas/ProjectResponse"
...
    PageAuditLogResponse:
      allOf:
        - $ref: "#/components/schemas/PageResponse"
        - type: object
          properties:
            content:
              type: array
              items:
                $ref: "#/components/schemas/AuditLogResponse"
security:
  - bearerAuth: []